dev-requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.8
# by the following command:
#
#    pip-compile '.\dev-requirements.in'
#
--extra-index-url https://download.pytorch.org/whl/cu113

asttokens==2.2.1
    # via stack-data
backcall==0.2.0
    # via ipython
black==23.3.0
    # via -r .\dev-requirements.in
build==0.10.0
    # via pip-tools
certifi==2023.5.7
    # via requests
charset-normalizer==3.1.0
    # via requests
click==8.1.3
    # via
    #   black
    #   pip-tools
colorama==0.4.6
    # via
    #   build
    #   click
    #   ipython
    #   tqdm
decorator==5.1.1
    # via ipython
diffusers==0.17.1
    # via -r .\requirements.in
executing==1.2.0
    # via stack-data
filelock==3.12.2
    # via
    #   diffusers
    #   huggingface-hub
    #   transformers
flake8==6.0.0
    # via -r .\dev-requirements.in
fsspec==2023.6.0
    # via huggingface-hub
huggingface-hub==0.15.1
    # via
    #   diffusers
    #   transformers
idna==3.4
    # via requests
importlib-metadata==6.7.0
    # via diffusers
ipython==8.12.2
    # via -r .\dev-requirements.in
isort==5.12.0
    # via -r .\dev-requirements.in
jedi==0.18.2
    # via ipython
matplotlib-inline==0.1.6
    # via ipython
mccabe==0.7.0
    # via flake8
mypy-extensions==1.0.0
    # via black
numpy==1.24.3
    # via
    #   diffusers
    #   torchvision
    #   transformers
packaging==23.1
    # via
    #   black
    #   build
    #   huggingface-hub
    #   transformers
parso==0.8.3
    # via jedi
pathspec==0.11.1
    # via black
pickleshare==0.7.5
    # via ipython
pillow==9.5.0
    # via
    #   diffusers
    #   torchvision
pip-tools==6.13.0
    # via -r .\dev-requirements.in
platformdirs==3.6.0
    # via black
prompt-toolkit==3.0.38
    # via ipython
pure-eval==0.2.2
    # via stack-data
pycodestyle==2.10.0
    # via flake8
pyflakes==3.0.1
    # via flake8
pygments==2.15.1
    # via ipython
pyproject-hooks==1.0.0
    # via build
pyyaml==6.0
    # via
    #   huggingface-hub
    #   transformers
regex==2023.6.3
    # via
    #   diffusers
    #   transformers
requests==2.31.0
    # via
    #   diffusers
    #   huggingface-hub
    #   torchvision
    #   transformers
safetensors==0.3.1
    # via transformers
six==1.16.0
    # via asttokens
stack-data==0.6.2
    # via ipython
tokenizers==0.13.3
    # via transformers
tomli==2.0.1
    # via
    #   black
    #   build
    #   pyproject-hooks
torch==1.12.1+cu113
    # via
    #   -r .\requirements.in
    #   torchvision
torchvision==0.13.1+cu113
    # via -r .\requirements.in
tqdm==4.65.0
    # via
    #   huggingface-hub
    #   transformers
traitlets==5.9.0
    # via
    #   ipython
    #   matplotlib-inline
transformers==4.30.2
    # via -r .\requirements.in
typing-extensions==4.6.3
    # via
    #   black
    #   huggingface-hub
    #   ipython
    #   torch
    #   torchvision
urllib3==2.0.3
    # via requests
wcwidth==0.2.6
    # via prompt-toolkit
wheel==0.40.0
    # via pip-tools
zipp==3.15.0
    # via importlib-metadata

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
